Eip712 module folder structure
