+++ eip712-starknet-account/src/client/index.ts
@@ -1,0 +1,8 @@
+import { MetaMaskInpageProvider } from '@metamask/providers';
+
+export class MetamaskClient {
+  constructor(protected provider: MetaMaskInpageProvider) {}
+
+  request = (method: string, ...params: unknown[]) =>
+    this.provider.request({ method, params });
+}
